refactor(a3p): deploy ymax1 persistently from g:ymax1's use.sh
Move ymax1's installAndStart out of an ava test action and into use.sh
(install-and-start.js), right after the wallet is provisioned and the
ContractControl invitation is redeemed. This makes the deployment persistent
across both a3p's build and test phases, giving the ymax1 vat a stable
vatID other proposals can reference — o:ymax1-multisig-control no longer
needs its own redundant deployment, and instead picks up the live kit via
getUpgradeKit.
