+++ a3p-integration/proposals/g:ymax1/consts.js
